Basic handling should be done. there is no strict error control however

+++ src/main/scala/app/models/conversionTables/LinearTable.scala
@@ -3,7 +3,7 @@ package app.models.conversionTables
 abstract class LinearTable(maxValue: Int) extends AsciiTable(maxValue) {
 
 	override def getTableValue(numValue: Int): Char = {
-		val resultIdx = ((_table.length - 1) * numValue / maxValue).floor.toInt
+		val resultIdx = ((_table.length - 1) * numValue / maxValue.toFloat).floor.toInt
 		_table.charAt(resultIdx)
 	}
 }
