+++ src/main/scala/app/models/conversionTables/NonLinearTable.scala
@@ -2,7 +2,7 @@ package app.models.conversionTables
 
 abstract class NonLinearTable(maxValue: Int) extends AsciiTable(maxValue) {
 	override def getTableValue(numValue: Int): Char = {
-		val resultIdx = _table.length.floor.toInt
+		val resultIdx = _table.length.toFloat.floor.toInt
 		_table.charAt(resultIdx)
 	}
 }
